requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
pywebpush==1.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
eventlet==0.35.2
python-engineio==4.9.0
python-socketio==5.11.0
